requirements.txt
# GridTrader Pro Requirements

# Core dependencies
python-telegram-bot==20.7
python-binance==1.0.19
python-dotenv==1.0.0

# Database
db-sqlite3

# Async support
asyncio

# Logging and utilities
pathlib

# Optional: For enhanced features
requests>=2.28.0
aiohttp>=3.8.0

# Development dependencies (optional)
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0
